+++ Services/Services.cs
@@ -343,6 +343,7 @@ public class AppSettings
     public string AppFont { get; set; } = "google-sans";
     public string CustomFont { get; set; } = "";
     public int FontSizeOffset { get; set; } = 0;
+    public int TaskbarHeight { get; set; } = 42;
     public List<string> ActiveCustomThemes { get; set; } = ["VRCNext v2 Preview"];
     public int GuiZoom { get; set; } = 100;
     public string DashBgPath { get; set; } = "";